+++ .github/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### New features
 
+* Added a new method to discretize Wigner functions that revolves Clenshaw summations. This method is expected to be fast and
+reliable for systems with high number of excitations, for which the pre-existing iterative method is known to be unstable. Users
+can select their preferred methods by setting the value of `Settings.DISCRETIZATION_METHOD` to either `interactive` (default) or
+`clenshaw`.
+
 * Added the `PhaseNoise(phase_stdev)` gate (non-Gaussian). Output is a mixed state in Fock representation.
   It is not based on a choi operator, but on a nonlinear transformation of the density matrix.
   [(#275)](https://github.com/XanaduAI/MrMustard/pull/275)
@@ -26,6 +31,7 @@
 [Filippo Miatto](https://github.com/ziofil), 
 [Samuele Ferracin](https://github.com/SamFerracin), [Yuan Yao](https://github.com/sylviemonet)
 
+---
 
 # Release 0.5.0 (current release)